Improved test setup for logger, improved test coverage for database
