+++ database/repository/script/script_test.go
@@ -1,7 +1,9 @@
-package tests
+package script
 
 import (
 	"fmt"
+	"io/ioutil"
+	"os"
 	"path/filepath"
 	"sync"
 	"testing"
@@ -10,17 +12,37 @@ import (
 	"github.com/thrasher-corp/gocryptotrader/database"
 	"github.com/thrasher-corp/gocryptotrader/database/drivers"
 	"github.com/thrasher-corp/gocryptotrader/database/repository"
-	"github.com/thrasher-corp/gocryptotrader/database/repository/script"
+	"github.com/thrasher-corp/gocryptotrader/database/testhelpers"
 	"github.com/thrasher-corp/goose"
 	"github.com/volatiletech/null"
 )
+
+func TestMain(m *testing.M) {
+	var err error
+	testhelpers.PostgresTestDatabase = testhelpers.GetConnectionDetails()
+
+	testhelpers.TempDir, err = ioutil.TempDir("", "gct-temp")
+	if err != nil {
+		fmt.Printf("failed to create temp file: %v", err)
+		os.Exit(1)
+	}
+
+	t := m.Run()
+
+	err = os.RemoveAll(testhelpers.TempDir)
+	if err != nil {
+		fmt.Printf("Failed to remove temp db file: %v", err)
+	}
+
+	os.Exit(t)
+}
 
 func TestScript(t *testing.T) {
 	testCases := []struct {
 		name   string
 		config *database.Config
 		runner func()
-		closer func(t *testing.T, dbConn *database.Db) error
+		closer func(dbConn *database.Db) error
 		output interface{}
 	}{
 		{
@@ -31,12 +53,12 @@ func TestScript(t *testing.T) {
 			},
 
 			writeScript,
-			closeDatabase,
+			testhelpers.CloseDatabase,
 			nil,
 		},
 		{
 			"Postgres-Write",
-			postgresTestDatabase,
+			testhelpers.PostgresTestDatabase,
 			writeScript,
 			nil,
 			nil,
@@ -47,16 +69,16 @@ func TestScript(t *testing.T) {
 		test := tests
 
 		t.Run(test.name, func(t *testing.T) {
-			if !checkValidConfig(t, &test.config.ConnectionDetails) {
+			if !testhelpers.CheckValidConfig(&test.config.ConnectionDetails) {
 				t.Skip("database not configured skipping test")
 			}
 
-			dbConn, err := connectToDatabase(t, test.config)
+			dbConn, err := testhelpers.ConnectToDatabase(test.config)
 
 			if err != nil {
 				t.Fatal(err)
 			}
-			path := filepath.Join("..", "migrations")
+			path := filepath.Join("..", "..", "migrations")
 			err = goose.Run("up", dbConn.SQL, repository.GetSQLDialect(), path, "")
 			if err != nil {
 				t.Fatalf("failed to run migrations %v", err)
@@ -67,7 +89,7 @@ func TestScript(t *testing.T) {
 			}
 
 			if test.closer != nil {
-				err = test.closer(t, dbConn)
+				err = test.closer(dbConn)
 				if err != nil {
 					t.Log(err)
 				}
@@ -85,7 +107,7 @@ func writeScript() {
 			defer wg.Done()
 			test := fmt.Sprintf("test-%v", x)
 			var data null.Bytes
-			script.Event(test, test, test, data, test, test, time.Now())
+			Event(test, test, test, data, test, test, time.Now())
 		}(x)
 	}
 	wg.Wait()
